renamed hash package to auth
