+++ auth/hash.go
@@ -1,4 +1,4 @@
-package hash
+package auth
 
 import "golang.org/x/crypto/bcrypt"
