+++ auth/hash_test.go
@@ -1,4 +1,4 @@
-package hash
+package auth
 
 import (
 	"testing"
